Make Addr a u64
So far Addr was defined to be a usize. That is in line with Rust
specifying that usize has the size of a pointer. However, once we start
using the library in a cross compilation context (e.g., using it to
symbolize parts of a dump captured on a different machine), the local
pointer size has admittedly little meaning.
With this change we define Addr to be of type u64, which is in-line with
the representation of many of the formats that we support. Doing so
makes us more independent of the actual pointer size as it basically is
meant as an upper limit on current sizes (that may not be entirely true
with research architectures such as CHERI, but it is still a reasonable
and arguably better approximation for what we expect to support).

+++ src/normalize/normalizer.rs
@@ -252,7 +252,7 @@ mod tests {
         assert_eq!(syms.len(), 1);
         let sym = syms.first().unwrap();
 
-        let the_answer_addr = unsafe { mmap.as_ptr().add(sym.addr) };
+        let the_answer_addr = unsafe { mmap.as_ptr().add(sym.addr as usize) };
         // Now just double check that everything worked out and the function
         // is actually where it was meant to be.
         let the_answer_fn =
@@ -301,7 +301,7 @@ mod tests {
                 .unwrap();
 
             let elf_mmap = mmap
-                .constrain(so.data_offset..so.data_offset + so.data.len())
+                .constrain(so.data_offset..so.data_offset + so.data.len() as u64)
                 .unwrap();
 
             // Look up the address of the `the_answer` function inside of the shared
@@ -316,7 +316,7 @@ mod tests {
             assert_eq!(syms.len(), 1);
             let sym = syms.first().unwrap();
 
-            let the_answer_addr = unsafe { elf_mmap.as_ptr().add(sym.addr) };
+            let the_answer_addr = unsafe { elf_mmap.as_ptr().add(sym.addr as usize) };
             // Now just double check that everything worked out and the function
             // is actually where it was meant to be.
             let the_answer_fn =
